+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/utils/test_merge_utils.py
@@ -61,6 +61,48 @@ class TestCollapseJobRoleEstimatesToPublishedLabels:
         )
 
 
+JOB_ROLE_ESTIMATES_SCHEMA_OVERRIDES = {
+    IndCQC.primary_service_type: CatColType.PrimaryServiceEnumType,
+    SLVCols.job_role_label: CatColType.PublishedJobRoleLabelEnumType,
+    METRIC: pl.Float64,
+}
+
+EXPECTED_SCHEMA_OVERRIDES = {
+    **JOB_ROLE_ESTIMATES_SCHEMA_OVERRIDES,
+    SLVCols.filled_posts_perm: pl.Float64,
+    SLVCols.filled_posts_temp: pl.Float64,
+    SLVCols.filled_posts_bank_or_pool: pl.Float64,
+    SLVCols.filled_posts_agency: pl.Float64,
+    SLVCols.filled_posts_other: pl.Float64,
+    SLVCols.employment_status_estimate_error: pl.Float64,
+}
+
+
 class TestApplyEmploymentStatusMagicNumbers:
-    def test_apply_employment_status_magic_numbers(self):
-        pass
+    @pytest.mark.parametrize(
+        "case",
+        [
+            pytest.param(case, id=case.id)
+            for case in Data.apply_employment_status_magic_numbers_test_cases
+        ],
+    )
+    def test_splits_metric_and_computes_error_column_as_expected(self, case):
+        job_role_estimates_lf = pl.LazyFrame(
+            case.job_role_estimates_data,
+            schema_overrides=JOB_ROLE_ESTIMATES_SCHEMA_OVERRIDES,
+        )
+        employment_status_rates_lf = pl.LazyFrame(case.employment_status_rates_data)
+        expected_lf = pl.LazyFrame(
+            case.expected_data, schema_overrides=EXPECTED_SCHEMA_OVERRIDES
+        )
+
+        returned_lf = job.apply_employment_status_magic_numbers(
+            job_role_estimates_lf, employment_status_rates_lf
+        )
+
+        pl_testing.assert_frame_equal(
+            returned_lf,
+            expected_lf,
+            check_row_order=False,
+            check_column_order=False,
+        )